Refactoring, updating traits
