+++ looker/Cargo.toml
@@ -3,16 +3,15 @@ name = "looker"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-starduck = { git = "https://github.com/ChipDepot/StarDuck.git", branch = "main" }
+# starduck = { git = "https://github.com/ChipDepot/StarDuck.git", branch = "main" }
+starduck = { path = "../../StarDuck/starduck" }
 
 project-root = "0.2.2"
 dotenv = "0.15"
 paho-mqtt = "0.12.2"
 
-redis = "0.23"
+redis = "0.24"
 
 log = "0.4.20"
 tracing = "0.1.37"
@@ -23,7 +22,9 @@ serde = "1"
 serde_json = "1.0"
 
 tokio = { version = "1.32", features = ["full"] }
-axum = { version = "0.6.20", features = ["tracing", "json"] }
+axum = { version = "0.7.2", features = ["tracing", "json"] }
 
 reqwest = { version = "0.11.22", features = ["json"] }
 url = "2.4.1"
+uuid = { version = "1.6.1", features = ["serde"] }
+chrono = { version = "0.4.31", features = ["serde"] }
