+++ looker/src/app/traits.rs
@@ -1,3 +1,5 @@
+use paho_mqtt::Message;
+
 pub trait Processor {
-    fn process_message(&mut self, message: &str);
+    fn process_message(&mut self, msg: Message);
 }
